feat(logic_compile): a Prolog compiler emitting C++, CUDA and Triton, at three integer widths
Rather than a JIT, the compiler emits SOURCE — C++23, CUDA for nvcc, and a Triton
kernel — so compiled predicates link into ordinary libraries and are read, diffed
and debugged like any other code in the tree.

WHAT IT COMPILES. Integer, moded, deterministic Prolog: `is/2`, the arithmetic
comparisons, `true`/`fail`/`!`, and calls to other compilable predicates. Modes are
GIVEN, never inferred, because a wrong mode changes what a program means rather
than how fast it runs. Anything outside that — a compound term, an unbound output
no goal binds, `findall`, a database update — is refused whole. A header holding
half a program is a trap, so the compiler never emits one.

THREE WIDTHS, and the combinations that cannot be honoured are refused up front:

  * bits64    — std::int64_t, every operation PRE-checked, because a guard written
                as `a + b < a` has already performed the undefined behaviour it
                meant to detect.
  * bits128   — __int128 via the compiler's overflow builtins, the same three
                nimblecas.int128 uses. Still BOUNDED: leaving the range is an
                overflow, not a wrong answer.
  * arbitrary — nimblecas::BigInt, where +, - and * CANNOT fail, so the emitted
                helpers do not pretend to check them. BigInt is a named-module
                entity with no header form, so this width emits a module-importing
                translation unit and says so in its own text.

  Triton takes bits64 only — `tl` has no int128 and no bignum dtype. CUDA refuses
  arbitrary precision outright: BigInt allocates per operation and device code
  cannot. Both refusals beat emitting something that silently wraps.

TAIL-CALL OPTIMISATION. A clause whose last goal calls its own predicate, with the
call's outputs being the head's outputs, is compiled to a LOOP rather than a call.
New arguments are computed into temporaries before being written back, since the
expressions read the arguments they are about to replace. An accumulator's
recursion depth stops being bounded by the native stack.

Repeated head variables now compile instead of being refused. `count(0, A, A)` is
the accumulator base clause and `p(X, X)` an equality test; with modes known,
neither needs unification — the first input occurrence defines the variable, a
later input occurrence tests against it, and an output occurrence assigns it.
Refusing these ruled out the single most common shape in the subset.

The emitted arithmetic is overflow-checked at every width. A compiler that turns a
checked interpreter into unchecked native code has not made the program faster, it
has made it wrong.

Verified by generating, compiling with clang and RUNNING: an accumulator over
3,000,000 tail iterations gives n(n+1)/2 at both 64 and 128 bits — which only
terminates because the tail call became a loop — and the arbitrary-precision build
sums to 500000500000 and represents 1000^40 at 121 digits. Full suite 165/165.

+++ src/gpu/gpu_bridge.h
@@ -9,6 +9,11 @@
 #ifndef NIMBLECAS_GPU_BRIDGE_H
 #define NIMBLECAS_GPU_BRIDGE_H
 
+// uint64_t rather than "unsigned long long": on LP64 the two are the same WIDTH but distinct
+// TYPES, and the module side speaks std::uint64_t. Naming the fixed-width type on both sides is
+// what makes the declaration and the definition actually match.
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -24,9 +29,9 @@ int nimblecas_gpu_device_count(void);
 // out_words must hold goal_count of them. The CPU path in nimblecas.logic_index is
 // AUTHORITATIVE; this kernel must produce the same bits. Returns 0 on success, or a non-zero
 // CUDA error code on failure.
-int nimblecas_gpu_index_probe_batch(const unsigned long long* clause_keys, int clause_count,
-                                    const unsigned long long* goal_keys, int goal_count,
-                                    unsigned long long* out_words);
+int nimblecas_gpu_index_probe_batch(const uint64_t* clause_keys, int clause_count,
+                                    const uint64_t* goal_keys, int goal_count,
+                                    uint64_t* out_words);
 
 // Evaluate the polynomial coeffs[0..n_coeffs-1] (low degree first) at each of the n
 // points x[0..n-1], writing p(x_i) to out[i]. Returns 0 on success, or a non-zero CUDA
